Move test container to other class
