+++ order-service/src/main/java/dev/shrkptv/orderservice/client/UserServiceClient.java
@@ -6,7 +6,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestParam;
 
-@FeignClient(name = "user-service", url = "http://localhost:8081/api/users")
+@FeignClient(name = "user-service", url = "${user-service.url")
 public interface UserServiceClient {
 
     @GetMapping("/email")
