+++ order-service/src/test/java/dev/shrkptv/orderservice/integration/AbstractIT.java
@@ -1,0 +1,17 @@
+package dev.shrkptv.orderservice.integration;
+
+import org.springframework.boot.test.context.SpringBootTest;
+import org.testcontainers.containers.PostgreSQLContainer;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
+
+@Testcontainers
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+public abstract class AbstractIT {
+
+    @Container
+    static PostgreSQLContainer<?> postgreSQLContainer = new PostgreSQLContainer<>("postgres:latest")
+            .withDatabaseName("order_service")
+            .withUsername("test")
+            .withPassword("test");
+}
